+++ finagle-http/src/test/scala/com/twitter/finagle/http/TraceInitializationTest.scala
@@ -1,6 +1,6 @@
 package com.twitter.finagle.http
 
-import com.twitter.finagle.builder.{ClientBuilder, ServerBuilder}
+import com.twitter.finagle.builder.ClientBuilder
 import com.twitter.finagle.tracing._
 import com.twitter.finagle.Service
 import com.twitter.util.{Await, Closable, Future}
@@ -72,12 +72,10 @@ class TraceInitializationTest extends AnyFunSuite {
   test("TraceId is propagated through the protocol (builder)") {
     import com.twitter.finagle
     testTraces { (serverTracer, clientTracer) =>
-      val server = ServerBuilder()
-        .name("theServer")
-        .bindTo(new InetSocketAddress(InetAddress.getLoopbackAddress, 0))
-        .stack(finagle.Http.server)
-        .tracer(serverTracer)
-        .build(Svc)
+      val server = finagle.Http.server
+        .withLabel("theServer")
+        .withTracer(serverTracer)
+        .serve(new InetSocketAddress(InetAddress.getLoopbackAddress, 0), Svc)
 
       val port = server.boundAddress.asInstanceOf[InetSocketAddress].getPort
       val client = ClientBuilder()